Exterior Penalty Method
Method - The exterior penalty method works by transforming the original problem into an unconstrained problem by converting the constraints into a penalty function term added to the original function. First, the constraints are all written as a collection of x1, x2, and constants less than 0 and labeled g1 through g4. The penalty function I used was of the form Ce^t where C was a small constant value and t was the left hand side of the constraints. This particular penalty function was used because if the constraint was negative or about zero, then the output would be zero. Whereas, if the constraint was positive, meaning the current point violated the constraint and was not in the feasible region, the penalty function would return a large value thereby penalizing the function we are trying to minimize. Next, the unconstrained function was formed by adding the original function to weighted penalty terms where the weights exponentially increased starting with 1, 10, 100, etc. Next the unconstrained function was minimized using Newton’s algorithm. After each iteration, the weights would be increased which further encourages the algorithm to choose points in the feasible region to minimize the function.