Interior Penalty Method
This method was similar to the exterior penalty method since the original problem was changed to an unconstrained optimization problem by converting the constraints into penalty terms. The main difference is the penalty function used was an inverse barrier function that kept the search within the feasible region instead of starting outside the feasible region and progressing inwards. To form the penalty term, the original constraints were written as the sum of all the components less than zero then the inverse of the constraint was taken. Ex. x<=0 was written as -1/x for the penalty term. In the unconstrained function the penalty terms were weighted starting at 1 then increased exponentially after each iteration from 1,10,100,etc. This method works because it deters the algorithm from surpassing the boundary by increasing the penalty term to a large number as the boundary is approached thereby pushing the search back inwards. The unconstrained function is minimized using Newton’s algorithm. In my runs the algorithm approaches the minimum but is repelled slightly since it’s near the boundary.