Zoutendijk's Method
The constraints are evaluated at the starting points then the search direction is set depending on if the constraints are satisfied or not. If all constraints are satisfied and the point is within the feasible region, the search direction is set to the negative of the gradient. If the point is outside the feasible region, the search direction is set to push the algorithm towards the feasible region by solving a set of equation to minimize the difference between the inner product of the search direction and the gradient of the function and the constraints separately. The point is updated by a line search where a step of size alpha is taken along the chosen search direction. The function is evaluated at the new point and iterations stop when the steps get really small and the function is changed negligibly per iteration.