Conjugate Gradient Method
The conjugate gradient code starts with setting the initial values of the x vector, matrix A, and vector b. The initial direction and alpha is set to the zero vector. Later on when the direction is not null, the alpha is an exact line search along direction d. The iterations start with calculating the new r value which is the residual of the difference between Ax and b and is a set of orthogonal vectors. The beta (B) value is calculated next using the recent and previous r vectors. The direction is updated using the r vector, beta value, and previous direction before calculating the new alpha length. This generates the d vector which is conjugate with respect to matrix A. The x vector is updated using the direction and alpha length. Iterations stopped when the L-2 norm of the gradient evaluated at the current x vector was less than the tolerance set at 10E-8.