Nonlinear conjugate gradient method
Similar to the conjugate gradient method, the nonlinear method computes d as a conjugate direction using the residuals for the r vector. In order to generalize, the nonlinear method uses the function’s gradient to find the direction and a step length (alpha) to perform a line search. The steepest descent direction is set to the negative of the gradient and is updated when no further progress is being made. The conjugate direction d is updated every iteration and depends on the gradient. After the line search is performed by calling the Goldstein-Armijo function, x is updated by adding the product of the step length and conjugate direction.