BFGS quasi-newton algorithm
Minimize the function given using the BFGS quasi-newton algorithm starting with the identify matrix as the initial Hessian approximation (J). 
Method: The BFGS method starts by setting the direction of descent as the negative of the gradient at the initial point. The step length was determined by line search using the Goldstein-Armijo criteria and the initial H matrix was set to the inner product of the identity matrix. For each iteration the gradient and direction of descent is updated by taking the result of multiplying matrix H by the gradient where H is the approximation to the Hessian matrix. After the direction is set, the line search returns the step length and is used to update x. Each iteration matrix H, which is the approximate Hessian at step i, is updated by adding two symmetric matrices which results in a positive definite matrix. The stopping method used was a combination of meeting the tolerance and reaching the maximum number of iterations. For the stopping criteria, if the number norm(gradient) / (1 + |f(x)|) was less than the tolerance then the algorithm was considered close enough to the minimum.
As the value of c was raised from 1 to 10 to 100 the algorithm performed poorer. The condition number of the Hessian is based on the ratio of the smallest to largest eigenvalue of the Hessian. When c=100, the condition number is larger which means the smallest eigenvalue is orders of magnitude smaller than the larger one. When c=1 and c=10, the eigenvalues are about the same order of magnitude so the condition number is smaller meaning the algorithm will run faster for that function.