Nelder-Mead Algorithm
Using the Nelder-Mead method, find the minimum of the given function starting with an initial simplex and setting up a triangle of points to search the area within the 10-radius constraint.
Method: The first part of the algorithm, initializing and beginning the iteration, involves calculating the function value of the simplex points and identifying the highest and lowest. Then the centroid is calculated between the maximum and minimum points before the reflection point is found. Here, and at a few other points in the code, the new point is checked to make sure it’s within the boundaries and if it falls outside the circle it is pushed back in by adding/subtracting .5 to each component of the x vector. If the reflection point does better than the worst simplex point but worse than the best point, the reflection is set to replace the worst point. However, if the reflection point is better than the best point, the algorithm gives it a chance to expand twice before replacing one of the original points. If the reflection point is worse than the maximum point, the algorithm contracts before replacing the worst point. The iterations stopped when q was smaller than the tolerance. Q is computed by calculating the average difference between the function values at each of the points in the simplex and the function value of the centroid then taking the square root. 
In the first function, the true minimum lies outside the boundaries, so the algorithm moves toward the point and gets one corner on the boundary for a few iterations. As you can see in the below snapshot of the simplex iterations, the algorithm nears the boundary and reflects outside the boundary as seen in the  transition from iter A to iter B. From here it moves along the boundary a little bit before collapsing in on the final simplex. The closest point on the circle to the real minimum is (-4.6483, 8.854) which is fairly close to the result the algorithm gives. There were two local minimum and the algorithm successfully found one given the choice for the starting simplex.